report number missed frame

+++ AOloopControl_IOtools.h
@@ -51,7 +51,7 @@ static void *compute_function_imtotal( void *ptr );
 static void *compute_function_dark_subtract( void *ptr );
 
 /** @brief Read image from WFS camera */
-int_fast8_t Read_cam_frame(long loop, int RM, int normalize, int PixelStreamMode, int InitSem);
+int Read_cam_frame(long loop, int RM, int normalize, int PixelStreamMode, int InitSem);
